Be more relaxed about inconsistencies of results
If issues requiring a PHP version greater than specified are found, but the general level info says it's OK, they probably are nothing to worry about e.g. dealt with proper function_exists calls. This behavior of the library does not seem to be really consistent when dealing with PHP versions < 5.2, but for our purposes it seems fine.